fix a variety of LSP issues
