+++ src/Directz.cpp.hpp
@@ -1,6 +1,5 @@
 #pragma once
-#include <dz/Shader.hpp>
-using namespace dz;
+#include <DirectZ.hpp>
 #include <atomic>
 #include <algorithm>
 #include <map>
@@ -40,13 +39,16 @@ using namespace dz;
 #endif
 #undef min
 #undef max
+#include "WindowImpl.hpp"
+#include "RendererImpl.hpp"
 namespace dz {
-	#include "WindowImpl.hpp"
-	#include "RendererImpl.hpp"
     /**
     * @brief Creates a Window given a Serial interface
     */
     WINDOW* window_create_from_serial(Serial& serial);
+
+    DirectRegistry* make_direct_registry();
+    void free_direct_registry();
 
     void set_env(const std::string& key, const std::string& value);
     std::string get_env(const std::string& key);
@@ -144,11 +146,47 @@ namespace dz
     void buffer_group_make_gpu_buffer(const std::string& name, ShaderBuffer& buffer);
     std::vector<WINDOW*>::iterator dr_get_windows_begin();
     std::vector<WINDOW*>::iterator dr_get_windows_end();
-    std::vector<WindowReflectableGroup*>::iterator dr_get_window_reflectable_entries_begin();
-    std::vector<WindowReflectableGroup*>::iterator dr_get_window_reflectable_entries_end();
     bool vk_check(const char* fn, VkResult result);
     void vk_log(const char* fn, VkResult result);
     bool recreate_swap_chain(Renderer* renderer);
+
+	struct QueueFamilyIndices
+	{
+		int32_t graphicsAndComputeFamily = -1;
+		int32_t presentFamily = -1;
+		bool isComplete() { return graphicsAndComputeFamily > -1 && presentFamily > -1; };
+	};
+	struct SwapChainSupportDetails
+	{
+		VkSurfaceCapabilitiesKHR capabilities;
+		std::vector<VkSurfaceFormatKHR> formats;
+		std::vector<VkPresentModeKHR> presentModes;
+	};
+    
+	VkSampleCountFlagBits get_max_usable_sample_count(DirectRegistry* direct_registry, Renderer* renderer);
+	void direct_registry_ensure_physical_device(DirectRegistry* direct_registry, Renderer* renderer);
+	uint32_t rate_device_suitability(DirectRegistry* direct_registry, Renderer* renderer, VkPhysicalDevice device);
+	bool is_device_suitable(DirectRegistry* direct_registry, Renderer* renderer, VkPhysicalDevice device);
+	QueueFamilyIndices find_queue_families(DirectRegistry* direct_registry, Renderer* renderer, VkPhysicalDevice device);
+	void direct_registry_ensure_logical_device(DirectRegistry* direct_registry, Renderer* renderer);
+	SwapChainSupportDetails query_swap_chain_support(Renderer* renderer, VkPhysicalDevice device);
+	VkSurfaceFormatKHR choose_swap_surface_format(const std::vector<VkSurfaceFormatKHR>& availableFormats);
+	VkPresentModeKHR choose_swap_present_mode(Renderer* renderer, const std::vector<VkPresentModeKHR>& availablePresentModes);
+	VkExtent2D choose_swap_extent(Renderer* renderer, VkSurfaceCapabilitiesKHR capabilities);
+	void ensure_command_pool(Renderer* renderer);
+	void ensure_command_buffers(Renderer* renderer);
+	void ensure_render_pass(Renderer* renderer);
+	void create_sync_objects(Renderer* renderer);
+	void pre_begin_render_pass(Renderer* renderer);
+	void begin_render_pass(Renderer* renderer);
+	void post_render_pass(Renderer* renderer);
+	bool swap_buffers(Renderer* renderer);
+	void renderer_draw_commands(Renderer* renderer, Shader* shader, const std::vector<DrawIndirectCommand>& commands);
+	void renderer_destroy(Renderer* renderer);
+	void destroy_swap_chain(Renderer* renderer);
+	void createBuffer(Renderer* renderer,
+		VkDeviceSize size, VkBufferUsageFlags usage, VkMemoryPropertyFlags properties,
+		VkBuffer& buffer, VkDeviceMemory& bufferMemory);
 }
 extern "C" DirectRegistry* dr_ptr;
 extern "C" DirectRegistry& dr;
