+++ src/Window.cpp
@@ -125,13 +125,12 @@ namespace dz {
 
 		if (dr.window_ptrs.size() == 1) {
 			ImGuiViewport* main_viewport = ImGui::GetMainViewport();
-			main_viewport->PlatformUserData = main_viewport->PlatformHandleRaw = window_get_native_handle(window);
+			main_viewport->PlatformHandleRaw = window_get_native_handle(window);
 			main_viewport->PlatformHandle = window;
 			main_viewport->Pos.x = window->x;
 			main_viewport->Pos.y = window->y;
 			main_viewport->Size.x = width;
 			main_viewport->Size.y = height;
-			main_viewport->PlatformWindowCreated = false;
 			window->imguiViewport = main_viewport;
 #ifdef _WIN32
 			dr.hwnd_root = window->hwnd;
@@ -188,9 +187,6 @@ namespace dz {
 			auto& vp = *window->imguiViewport;
             vp.PlatformHandle = nullptr;
             vp.PlatformHandleRaw = nullptr;
-			if (!destroy_remaining)
-	            vp.RendererUserData = nullptr;
-            vp.PlatformUserData = nullptr;
 		}
 		window->destroy_platform();
 		auto& event_interface = *window->event_interface; 
@@ -435,15 +431,16 @@ namespace dz {
 			setDPIAware = true;
 		}
 		hInstance = GetModuleHandle(NULL);
-		WNDCLASSEX wc = {0};
+		WNDCLASSEXW wc = {0};
 		// wc.cbSize = sizeof(WNDCLASS);
-		wc.cbSize = sizeof(WNDCLASSEX);
+		wc.cbSize = sizeof(WNDCLASSEXW);
 		wc.style = CS_VREDRAW | CS_HREDRAW;
 		wc.lpfnWndProc = wndproc;
 		wc.hInstance = hInstance;
-		wc.lpszClassName = title.c_str();
+		auto wtitle = string_to_wstring(title);
+		wc.lpszClassName = wtitle.c_str();
 		wc.hCursor = LoadCursor(NULL, IDC_ARROW);
-		RegisterClassEx(&wc);
+		RegisterClassExW(&wc);
 		dpiScale = 1.0f;
 		HDC screen = GetDC(NULL);
 		int32_t dpi = GetDeviceCaps(screen, LOGPIXELSX);
@@ -456,10 +453,10 @@ namespace dz {
 		AdjustWindowRectEx(&desiredRect, wsStyle, FALSE, exStyle);
 		adjustedWidth = desiredRect.right - desiredRect.left;
 		adjustedHeight = desiredRect.bottom - desiredRect.top;
-		hwnd = CreateWindowEx(
+		hwnd = CreateWindowExW(
 			exStyle,
-			title.c_str(),
-			title.c_str(),
+			wtitle.c_str(),
+			wtitle.c_str(),
 			wsStyle,
 			x == -1 ? CW_USEDEFAULT : x,
 			y == -1 ? CW_USEDEFAULT : y,
