+++ src/runtime.cpp
@@ -1,3 +1,4 @@
 #include <DirectZ.hpp>
-DirectRegistry* dr_ptr = make_direct_registry();
+#include "Directz.cpp.hpp"
+DirectRegistry* dr_ptr = dz::make_direct_registry();
 DirectRegistry& dr = *dr_ptr;
